+++ app/src/main/java/com/korsun/sunrise/api/response/Main.java
@@ -5,8 +5,8 @@ package com.korsun.sunrise.api.response;
  */
 public final class Main {
     public float temp = 0;
-    public int pressure = 0;
-    public int humidity = 0;
+    public float pressure = 0;
+    public float humidity = 0;
     public float temp_min = 0;
     public float temp_max = 0;
 }
